Task-372: WIP adding api endpoints and connecting to logic
